requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
streamlit
requests
pandas
fpdf2
xlsxwriter
streamlit-calendar
openpyxl